+++ simulator/profiling/attention/attention_wrapper.py
@@ -3,6 +3,12 @@ from typing import List
 
 import numpy as np
 import torch
+
+from simulator.profiling.common.cuda_timer import CudaTimer
+import sarathi.metrics.cuda_timer
+# monkey patching the CudaTimer class to use the sarathi implementation
+sarathi.metrics.cuda_timer.CudaTimer = CudaTimer
+
 from sarathi.model_executor.attention import (
     AttentionBackend,
     get_attention_wrapper,
@@ -27,10 +33,8 @@ class AttentionWrapper:
         max_model_len: int,
         block_size: int,
         attention_backend: AttentionBackend,
-        profile_method: ProfileMethod,
     ):
-        self._profile_method = profile_method
-        self.time_stats_store = TimerStatsStore(profile_method=profile_method)
+        self.time_stats_store = TimerStatsStore(profile_method="kineto")
 
         self._n_embd = model_config.embedding_dim
         self._n_q_head = model_config.num_q_heads
@@ -58,14 +62,14 @@ class AttentionWrapper:
             self._block_size,
             self._device,
         )
-        self._blocks_per_sequence = ceil(max_model_len / self._block_size)
+        self._max_blocks_per_sequence = ceil(max_model_len / self._block_size)
 
     def _get_input_tensors(
         self,
         attention_input: AttentionInput,
     ):
         total_num_blocks = max(
-            10000, 1 + attention_input.batch_size * self._blocks_per_sequence
+            10000, 1 + attention_input.batch_size * self._max_blocks_per_sequence
         )
         num_tokens_per_seq = (
             attention_input.prefill_chunk_size if attention_input.is_prefill else 1
@@ -97,12 +101,13 @@ class AttentionWrapper:
         # Create SequenceMetadataProxy objects corresponding to AttentionInput
         seq_metadata_list: List[SequenceMetadataProxy] = []
         for _ in range(attention_input.batch_size):
+            num_blocks = ceil((num_tokens_per_seq + attention_input.kv_cache_size) / self._block_size)
             seq_metadata = SequenceMetadataProxy(
                 is_prompt=attention_input.is_prefill,
                 total_len=num_tokens_per_seq + attention_input.kv_cache_size,
                 processed_len=attention_input.kv_cache_size,
                 block_table=np.random.default_rng()
-                .integers(low=0, high=total_num_blocks, size=self._blocks_per_sequence)
+                .integers(low=0, high=total_num_blocks, size=num_blocks)
                 .tolist(),
             )
             seq_metadata_list.append(seq_metadata)
@@ -150,5 +155,4 @@ class AttentionWrapper:
             "kv_cache_size": attention_input.kv_cache_size,
             "is_prefill": attention_input.is_prefill,
             "attention_backend": self._attention_backend,
-            "profile_method": self._profile_method,
         }
