+++ simulator/profiling/attention/main.py
@@ -12,7 +12,7 @@ from tqdm import tqdm
 from simulator.profiling.attention.attention_input import AttentionInput
 from simulator.profiling.attention.attention_wrapper import AttentionWrapper
 from simulator.profiling.common.model_config import ModelConfig
-from simulator.profiling.utils import ProfileMethod, get_attention_input_combinations
+from simulator.profiling.utils import get_attention_input_combinations
 
 
 def parse_args():
@@ -25,7 +25,7 @@ def parse_args():
     parser.add_argument(
         "--num_gpus",
         type=int,
-        default=8,
+        default=4,
         help="Number of GPUs to use for profiling",
     )
     parser.add_argument(
@@ -39,8 +39,7 @@ def parse_args():
         type=str,
         nargs="+",
         default=[
-            "microsoft/phi-2",
-            "internlm/internlm-20b",
+            # "internlm/internlm-20b",
             "Qwen/Qwen-72B",
             "meta-llama/Llama-2-7b-hf",
             "codellama/CodeLlama-34b-Instruct-hf",
@@ -90,18 +89,8 @@ def parse_args():
         help="Only profile the prefill",
     )
     parser.add_argument(
-        "--profile_method",
-        default=ProfileMethod.KINETO,
-        const=ProfileMethod.KINETO,
-        nargs="?",
-        choices=[e.value for e in ProfileMethod],
-        help="Method to use for measuring time taken by operations (default: %(default)s)",
-    )
-    parser.add_argument(
         "--attention_backend",
         default=AttentionBackend.FLASH_ATTENTION,
-        const=AttentionBackend.FLASH_ATTENTION,
-        nargs="?",
         choices=[e.value for e in AttentionBackend],
         help="The attention backend to profile (default: %(default)s)",
     )
@@ -126,9 +115,6 @@ def profile_model(
     pbar: Any,
 ):
     model_config = ModelConfig.from_model_name(model)
-    output_file_path = f"{args.output_dir}/{model}.csv"
-    # create parent directory if it doesn't exist
-    os.makedirs(os.path.dirname(output_file_path), exist_ok=True)
 
     promises = []
     all_results = []
@@ -148,7 +134,6 @@ def profile_model(
                 args.max_model_len,
                 args.block_size,
                 args.attention_backend,
-                args.profile_method,
             )
             for _ in range(args.num_gpus)
         ]
@@ -178,10 +163,6 @@ def profile_model(
         .add_prefix("time_stats.")
         .join(df.drop(columns=["time_stats"]))
     )
-    # write results to a csv file
-    df.to_csv(output_file_path, index=False)
-    df.to_json(f"{args.output_dir}/{model}.json", orient="records")
-
     return df
 
 
@@ -204,8 +185,6 @@ def main():
 
     pbar = tqdm(total=len(list(total_combos)))
 
-    result_dfs = []
-
     for model in args.models:
         result_df = profile_model(
             args,
@@ -213,12 +192,9 @@ def main():
             input_combinations,
             pbar,
         )
-        result_dfs.append(result_df)
-
-    # combine all results into a single dataframe
-    combined_df = pd.concat(result_dfs, ignore_index=True)
-    # write combined results to a csv file
-    combined_df.to_csv(f"{args.output_dir}/combined.csv", index=False)
+        # model name would contain '/', so create a directory as required
+        os.makedirs(f"{args.output_dir}/{model}", exist_ok=True)
+        result_df.to_csv(f"{args.output_dir}/{model}/attention.csv", index=False)
 
 
 if __name__ == "__main__":
