+++ simulator/profiling/mlp/mlp_wrapper.py
@@ -1,14 +1,19 @@
-import json
 import os
 
-import numpy as np
 import torch
+
+from simulator.profiling.common.cuda_timer import CudaTimer
+import sarathi.metrics.cuda_timer
+# monkey patching the CudaTimer class to use the sarathi implementation
+sarathi.metrics.cuda_timer.CudaTimer = CudaTimer
+
 from sarathi.model_executor.weight_utils import initialize_dummy_weights
 
 from simulator.profiling.mlp.mlp_impl import GPTModel
 from simulator.profiling.common.model_config import ModelConfig
-from simulator.profiling.common.timer_stats_store import TimerStatsStore\
+from simulator.profiling.common.timer_stats_store import TimerStatsStore
 from simulator.profiling.utils import ProfileMethod
+from simulator.profiling.utils.record_function_tracer import RecordFunctionTracer
 
 WARMUP_STEPS = 2
 ACTIVE_STEPS = 20
@@ -53,6 +58,7 @@ class MlpWrapper:
             dtype=torch.long,
         )
         positions = torch.arange(num_tokens, device="cuda", dtype=torch.long)
+
         if self.profile_method == ProfileMethod.RECORD_FUNCTION:
             # Run the model once without capturing the graph.
             # This is to make sure that the captured graph does not include the
@@ -65,16 +71,15 @@ class MlpWrapper:
 
             self.timer_stats_store.clear_stats()
 
-            self.start_profiling("cuda")
-            self.model(
-                input_ids,
-                positions,
-            )
-            self.stop_profiling()
-            torch.cuda.synchronize()
-            model_hash = str(hash(self.model_config))[:10]
-            trace_path = f"{self.output_dir}/profiler_traces/profiler_trace_{model_hash}_num_tokens_{num_tokens}_tp_{self.num_tensor_parallel_workers}.json"
-            self.profiler.export_chrome_trace(trace_path)
+            record_function_tracer = RecordFunctionTracer(self.output_dir)
+
+            with record_function_tracer:
+                self.model(
+                    input_ids,
+                    positions,
+                )
+            
+            time_stats = record_function_tracer.get_operation_time_stats()
         else:
             for _ in range(WARMUP_STEPS):
                 self.model(
@@ -94,12 +99,10 @@ class MlpWrapper:
 
             torch.cuda.synchronize()
 
+            time_stats = self.timer_stats_store.get_stats()
+
         stats = {
-            "time_stats": (
-                self.get_operation_time_stats(trace_path)
-                if self.profile_method == ProfileMethod.RECORD_FUNCTION
-                else self.timer_stats_store.get_stats()
-            ),
+            "time_stats": time_stats,
             "n_head": self.model_config.num_q_heads,
             "n_kv_head": self.model_config.num_kv_heads,
             "n_embd": self.model_config.embedding_dim,
@@ -113,90 +116,3 @@ class MlpWrapper:
 
         return stats
 
-    def start_profiling(self, activity) -> None:
-        if activity == "cpu":
-            activities = [
-                torch.profiler.ProfilerActivity.CPU,
-            ]
-        elif activity == "cuda":
-            activities = [
-                torch.profiler.ProfilerActivity.CPU,
-                torch.profiler.ProfilerActivity.CUDA,
-            ]
-        self.profiler = torch.profiler.profile(
-            activities=activities,
-        )
-        self.profiler.__enter__()
-
-    def stop_profiling(self) -> None:
-        self.profiler.__exit__(None, None, None)
-
-    def find_children(self, trace, event):
-        if not "dur" in event or not "ts" in event:
-            return
-
-        children = []
-        for e in trace:
-            if not "dur" in e or not "ts" in e:
-                continue
-
-            # if the ts of the child is completely within the ts of the parent
-            if (
-                e["ts"] > event["ts"]
-                and e["ts"] + e["dur"] < event["ts"] + event["dur"]
-            ):
-                children.append(e)
-        return children
-
-    def find_correlated_event(self, trace, event):
-        if not "args" in event or not "correlation" in event["args"]:
-            return
-
-        for e in trace:
-            if not "args" in e or not "correlation" in e["args"]:
-                continue
-
-            if e == event:
-                continue
-
-            if e["args"]["correlation"] == event["args"]["correlation"]:
-                return e
-        return
-
-    def get_operation_time_stats(self, trace_path: str):
-        # print("get_operation_time_stats")
-        stats = {}
-
-        trace = json.load(open(trace_path, "r"))["traceEvents"]
-        for event in trace:
-            if not "cat" in event or event["cat"] != "user_annotation":
-                continue
-            children = self.find_children(trace, event)
-            cuda_time = 0
-            for child in children:
-                if not "cat" in child or child["cat"] != "cuda_runtime":
-                    continue
-                correlated_event = self.find_correlated_event(trace, child)
-                if not correlated_event:
-                    continue
-                cuda_time += correlated_event["dur"]
-            if cuda_time == 0:
-                continue
-
-            name = event["name"].replace("vidur_", "")
-
-            if name not in stats:
-                stats[name] = []
-
-            stats[name].append(cuda_time * 1e-3)  # to convert to ms
-
-        return {
-            operation: {
-                "min": np.min(times),
-                "max": np.max(times),
-                "mean": np.mean(times),
-                "median": np.median(times),
-                "std": np.std(times),
-            }
-            for operation, times in stats.items()
-        }
